# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Multi-Tenant Project Management API",
        "description": "A multi-tenant project and task management API built with Laravel",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/t/{organization}/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List all projects",
                "operationId": "9de63729d6fabaedc5a5ec8469ba1adb",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "operationId": "ad7c88347562854e98d4f745aaca1127",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Project"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Project description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "planning",
                                            "in_progress",
                                            "on_hold",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "planning"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/t/{organization}/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a specific project",
                "operationId": "35c5a78180a73ddde5c288e8b58c7724",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "operationId": "04756a8d8cf5ea9d42d0132a602f90ae",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "planning",
                                            "in_progress",
                                            "on_hold",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "operationId": "aa6bc4b04adfe442fd07ce75e2ff4724",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Project": {
                "required": [
                    "id",
                    "name",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "My Project"
                    },
                    "description": {
                        "type": "string",
                        "example": "Project description",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "planning",
                            "in_progress",
                            "on_hold",
                            "completed",
                            "cancelled"
                        ],
                        "example": "planning"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}